+++ crest/rationalizers/lightning_models/highlights/transformers/base_matching.py
@@ -255,12 +255,12 @@ class TransformerBaseMatchingRationalizer(BaseRationalizer):
             # t5 encoder and decoder receives word_emb and a raw mask
             gen_h = self.ff_gen_encoder(
                 inputs_embeds=gen_e,
-                attention_mask=mask
+                attention_mask=mask.long()
             )
             if self.ff_gen_use_decoder:
                 gen_h = self.ff_gen_decoder(
                     inputs_embeds=gen_e,
-                    attention_mask=mask,
+                    attention_mask=mask.long(),
                     encoder_hidden_states=gen_h.last_hidden_state,
                 )
         else:
